# app/routes/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Content-X (Groot) API",
    "description": "Schema-driven CMS API for the Fynd Platform extension **groot**. Routes under `/api` are served through the FDK extension router (`fdk-extension-javascript`); clients must have a valid extension / platform session (cookies and OAuth flow). `company_id` and `application_id` are taken from session, query, or body per `request-context.helper.js`. Published CMS instance routes use `router.param('slug', loadDefinition)` and require a **published** definition for that slug.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the extension (paths include `/api/v1/...`)"
    }
  ],
...
  "paths": {
...
    "/api/v1/cms/definitions": {
      "get": {
        "tags": ["Definitions"],
        "summary": "List object definitions for the current company + application",
        "operationId": "listObjectDefinitions",
        "parameters": [
          {
            "name": "application_id",
            "in": "query",
            "description": "Target application (also from FDK session / body in some clients).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ObjectDefinition" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Definitions"],
        "summary": "Create object definition",
        "operationId": "createObjectDefinition",
        "parameters": [
          {
            "name": "application_id",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ObjectDefinitionCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectDefinition" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/api/v1/cms/{slug}": {
      "get": {
        "tags": ["CMS instances"],
        "summary": "List instances (standalone definitions only; 403 if embedded_only)",
        "operationId": "listObjectInstances",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          {
            "name": "page_size",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "default": "createdAt" } },
          {
            "name": "sort_order",
            "in": "query",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "Filter object; structure depends on definition filterable fields (passed through `buildFilterQuery`).",
            "schema": { "type": "string" },
            "examples": {
              "json": { "value": "{}" }
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InstanceListResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/DefinitionNotFound" }
        }
      },
      "post": {
        "tags": ["CMS instances"],
        "summary": "Create instance (multipart if file fields; 403 if embedded_only)",
        "operationId": "createObjectInstance",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "description": "JSON fields and/or multipart file parts named after `file` fields in the definition.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/DefinitionNotFound" }
        }
      }
    },
...
    "/api/v1/cms/{slug}/bulk/template": {
      "get": {
        "tags": ["CMS instances"],
        "summary": "Download bulk import template (CSV or XLSX)",
        "operationId": "downloadBulkTemplate",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["csv", "xlsx"], "default": "xlsx" }
          },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/csv": { "schema": { "type": "string", "format": "binary" } },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/DefinitionNotFound" }
        }
      }
    },
    "/api/v1/cms/{slug}/bulk/import": {
      "post": {
        "tags": ["CMS instances"],
        "summary": "Bulk import from spreadsheet upload",
        "operationId": "bulkImportInstances",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "CSV or XLSX" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkImportResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/DefinitionNotFound" }
        }
      }
    },
...
    "/api/v1/cms/{slug}/{id}": {
      "get": {
        "tags": ["CMS instances"],
        "summary": "Get instance by MongoDB id",
        "operationId": "getObjectInstance",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "description": "MongoDB ObjectId string" }
          },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["CMS instances"],
        "summary": "Replace/update instance",
        "operationId": "updateObjectInstance",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "additionalProperties": true }
            },
            "multipart/form-data": {
              "schema": { "type": "object", "additionalProperties": true }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["CMS instances"],
        "summary": "Delete instance",
        "operationId": "deleteObjectInstance",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "Slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "Published object definition slug (loads definition via `loadDefinition`).",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
...
      "FieldSchema": {
        "type": "object",
        "description": "Single field in an object definition",
        "properties": {
          "name": { "type": "string" },
          "label": { "type": "string" },
          "field_type": {
            "type": "string",
            "enum": [
              "text",
              "long_text",
              "number",
              "file",
              "date",
              "date_time",
              "url",
              "json",
              "html",
              "multiple_values",
              "dropdown",
              "checkbox",
              "radio",
              "product",
              "collection",
              "custom_object"
            ]
          },
          "required": { "type": "boolean" },
          "regex_pattern": { "type": "string" },
          "is_filterable": { "type": "boolean" },
          "options": { "type": "array", "items": { "type": "string" } },
          "default_value": {},
          "multiple": { "type": "boolean" },
          "relation": { "type": "string", "enum": ["none", "product", "collection"] },
          "placeholder": { "type": "string" },
          "ref_definition_slug": { "type": "string" },
          "file_config": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": true
      },
      "ObjectDefinition": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "company_id": { "type": "string" },
          "application_id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "usage_mode": { "type": "string", "enum": ["standalone", "embedded_only"] },
          "status": { "type": "string", "enum": ["DRAFT", "PUBLISHED"] },
          "version": { "type": "integer" },
          "fields": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FieldSchema" }
          },
          "published_at": { "type": "string", "format": "date-time" },
          "created_by": { "type": "string" },
          "updated_by": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": true
      },
      "ObjectDefinitionCreate": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string", "description": "Optional; derived from name if omitted" },
          "description": { "type": "string" },
          "usage_mode": { "type": "string", "enum": ["standalone", "embedded_only"] },
          "fields": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FieldSchema" }
          }
        }
      },
...
      "InstanceListResponse": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "items": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          }
        }
      },
      "BulkImportResult": {
        "type": "object",
        "properties": {
          "created_count": { "type": "integer" },
          "failed_count": { "type": "integer" },
          "created_ids": { "type": "array", "items": { "type": "string" } },
          "failed": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "row": { "type": "integer" },
                "message": { "type": "string" }
              }
            }
          },
          "message": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "DefinitionNotFound": {
        "description": "Published object definition not found for slug",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "ForbiddenEmbeddedOnly": {
        "description": "Definition is embedded_only; standalone instance APIs are disabled",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      }
    }
  }
}